+++ tests/test_sqlite_util.py
@@ -96,3 +96,23 @@ def test_create_table_as_sql_invalidparams(kwargs, expected_error):
 
     with pytest.raises(ValueError, match=expected_error):
         _sqlite_util.create_table_as_sql(**kwargs)
+
+
+def test_execute_sql(tmp_path):
+    test_path = test_helper.get_testfile(testfile="polygon-parcel", dst_dir=tmp_path)
+    assert gfo.has_spatial_index(test_path)
+    info_input = gfo.get_layerinfo(test_path)
+    nb_deleted = 0
+
+    # Execute one statement
+    sql_stmt = "DELETE FROM parcels WHERE rowid = (SELECT MIN(rowid) FROM parcels)"
+    _sqlite_util.execute_sql(test_path, sql_stmt=sql_stmt)
+    nb_deleted += 1
+    info = gfo.get_layerinfo(test_path)
+    assert info.featurecount == info_input.featurecount - nb_deleted
+
+    # Execute a list of statements
+    _sqlite_util.execute_sql(test_path, sql_stmt=[sql_stmt, sql_stmt])
+    nb_deleted += 2
+    info = gfo.get_layerinfo(test_path)
+    assert info.featurecount == info_input.featurecount - nb_deleted