+++ CHANGELOG.md
@@ -9,9 +9,13 @@
 - Improve performance of spatial operations when only one batch is used (#271)
 - Add checks that `output_path` must not be equal to the/an `input_path` for geo
   operations (#246)
+- Follow geopandas behaviour of using shapely2 and/or pygeos instead of forcing pygeos
+  (#294)
+- Many small improvements to logging, documentation, error messages,...
 
 ### Bugs fixed
 
+- Fix parameter `index` in `to_file` being ~ ignored (#285)
 - Fix "json" aggregate column handling in dissolve on line and point input files gives
   wrong results (#257)
 
@@ -19,6 +23,10 @@
 
 - `makevalid` parameter `precision` is renamed to `gridsize` as this is the typical
   terminology in other libraries (#273)
+- removed the long-deprecated functions `get_driver`, `get_driver_for_ext`,
+  `to_multi_type` and `to_generaltypeid`  (#276)
+- removed the long-deprecated `vector_util`, `geofileops.geofile` and
+  `geofileops.geofileops` namespaces (#276)
 
 ## 0.7.0 (2023-03-17)